Fixed commit
Removed Nugget autorestore
Changed clases for the .debugger
