+++ Tests/Yodii.Script.Debugger.Tests/Properties/AssemblyInfo.cs
