+++ Tests/Yodii.Script.Debugger.Tests/packages.config
