+++ Yodii.Script.Debugger/DynamicScopeDebugger.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Yodii.Script.Debugger
+{
+    class DynamicScopeDebugger : DynamicScope
+    {
+    }
+}
