+++ Yodii.Script.Debugger/Yodii.Script.Debugger.csproj
@@ -41,6 +41,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BreakableVisitor.cs" />
+    <Compile Include="DynamicScopeDebugger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -48,6 +49,9 @@
       <Project>{388870e4-cf4e-45b5-bbec-ec6fad2e7490}</Project>
       <Name>Yodii.Script</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="BreakpointManagerDebugger.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
